Rename regex.ts file to forms.ts
